open modal if no data given
