+++ client/src/actions/accounts.js
@@ -6,7 +6,7 @@ export const BALANCE_OF_CLN = createRequestTypes('BALANCE_OF_CLN')
 
 export const FETCH_BALANCES = createRequestTypes('FETCH_BALANCES')
 export const FETCH_TOKENS_WITH_BALANCES = createRequestTypes('FETCH_TOKENS_WITH_BALANCES')
-export const ADD_USER_INFORMATION = createRequestTypes('ADD_USER_INFORMATION')
+export const ADD_USER = createRequestTypes('ADD_USER')
 
 export const balanceOfToken = (tokenAddress, accountAddress) => action(BALANCE_OF_TOKEN.REQUEST, {tokenAddress, accountAddress})
 export const balanceOfNative = (accountAddress) => action(BALANCE_OF_NATIVE.REQUEST, {accountAddress})
