+++ client/src/actions/user.js
@@ -2,8 +2,12 @@ import {createRequestTypes, requestAction} from './utils'
 
 export const LOGIN = createRequestTypes('LOGIN')
 export const LOGOUT = createRequestTypes('LOGOUT')
-export const ADD_USER_INFORMATION = createRequestTypes('ADD_USER_INFORMATION')
+
+export const ADD_USER = createRequestTypes('ADD_USER')
+export const IS_USER_EXISTS = createRequestTypes('IS_USER_EXISTS')
 
 export const login = () => requestAction(LOGIN)
 export const logout = () => requestAction(LOGOUT)
-export const addUserInformation = (user) => requestAction(ADD_USER_INFORMATION, {user})
+
+export const addUser = (user, tokenAddress) => requestAction(ADD_USER, {user, tokenAddress})
+export const isUserExists = (accountAddress) => requestAction(IS_USER_EXISTS, {accountAddress})
