+++ client/src/reducers/screens/dashboard.js
@@ -1,12 +1,14 @@
 import {FETCH_TOKEN_STATISTICS} from 'actions/token'
-import {ADD_USER_INFORMATION} from 'actions/accounts'
+import {ADD_USER, IS_USER_EXISTS} from 'actions/user'
 
 export default (state = {}, action) => {
   switch (action.type) {
     case FETCH_TOKEN_STATISTICS.SUCCESS:
       return {...state, ...action.response}
-    case ADD_USER_INFORMATION.SUCCESS:
+    case ADD_USER.SUCCESS:
       return {...state, informationAdded: true}
+    case IS_USER_EXISTS.SUCCESS:
+      return {...state, ...action.response}
     default:
       return state
   }
